pyproject.toml
[project]
name = "writercompany"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi[standard]>=0.135.3",
    "google-genai>=1.73.1",
    "langgraph>=1.1.7",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn[standard]>=0.44.0",
]
